noVariableVariables and towards bleedingEdge https://github.com/phpstan/phpstan-strict-rules https://phpstan.org/blog/what-is-bleeding-edge

+++ lib/Minz/Extension.php
@@ -19,8 +19,6 @@ abstract class Minz_Extension {
 	private $version;
 	/** @var 'system'|'user' */
 	private $type;
-	/** @var string */
-	private $config_key = 'extensions';
 	/** @var array<string,mixed>|null */
 	private $user_configuration;
 	/** @var array<string,mixed>|null */
@@ -175,8 +173,11 @@ abstract class Minz_Extension {
 			$mtime = @filemtime("{$this->path}/static/{$filename}");
 		} else {
 			$username = Minz_User::name();
-			$path = USERS_PATH . "/{$username}/{$this->config_key}/{$this->getName()}/{$filename}";
-			$file_name_url = urlencode("{$username}/{$this->config_key}/{$this->getName()}/{$filename}");
+			if ($username == null) {
+				return '';
+			}
+			$path = USERS_PATH . "/{$username}/extensions/{$this->getName()}/{$filename}";
+			$file_name_url = urlencode("{$username}/extensions/{$this->getName()}/{$filename}");
 			$mtime = @filemtime($path);
 		}
 
@@ -242,27 +243,19 @@ abstract class Minz_Extension {
 				return false;
 		}
 
-		if (!$conf->hasParam($this->config_key)) {
+		if (!$conf->hasParam('extensions')) {
 			return false;
 		}
 
-		$extensions = $conf->{$this->config_key};
-		return array_key_exists($this->getName(), $extensions);
+		return array_key_exists($this->getName(), $conf->extensions);
 	}
 
 	/**
-	 * @phpstan-param 'system'|'user' $type
 	 * @return array<string,mixed>
 	 */
-	private function getConfiguration(string $type): array {
-		if (!$this->isConfigurationEnabled($type) || !$this->isExtensionConfigured($type)) {
-			return [];
-		}
-
-		if ($type === 'user') {
-			return FreshRSS_Context::userConf()->{$this->config_key}[$this->getName()];
-		} elseif ($type === 'system') {
-			return FreshRSS_Context::systemConf()->{$this->config_key}[$this->getName()];
+	public final function getSystemConfiguration(): array {
+		if ($this->isConfigurationEnabled('system') && $this->isExtensionConfigured('system')) {
+			return FreshRSS_Context::systemConf()->extensions[$this->getName()];
 		}
 		return [];
 	}
@@ -270,15 +263,11 @@ abstract class Minz_Extension {
 	/**
 	 * @return array<string,mixed>
 	 */
-	public final function getSystemConfiguration(): array {
-		return $this->getConfiguration('system');
-	}
-
-	/**
-	 * @return array<string,mixed>
-	 */
 	public final function getUserConfiguration(): array {
-		return $this->getConfiguration('user');
+		if ($this->isConfigurationEnabled('user') && $this->isExtensionConfigured('user')) {
+			return FreshRSS_Context::userConf()->extensions[$this->getName()];
+		}
+		return [];
 	}
 
 	/**
@@ -327,14 +316,14 @@ abstract class Minz_Extension {
 				return;
 		}
 
-		if ($conf->hasParam($this->config_key)) {
-			$extensions = $conf->{$this->config_key};
+		if ($conf->hasParam('extensions')) {
+			$extensions = $conf->extensions;
 		} else {
 			$extensions = [];
 		}
 		$extensions[$this->getName()] = $configuration;
 
-		$conf->{$this->config_key} = $extensions;
+		$conf->extensions = $extensions;
 		$conf->save();
 	}
 
@@ -367,13 +356,12 @@ abstract class Minz_Extension {
 				return;
 		}
 
-		$extensions = $conf->{$this->config_key};
+		$extensions = $conf->extensions;
 		unset($extensions[$this->getName()]);
 		if (empty($extensions)) {
-			$extensions = null;
+			$extensions = [];
 		}
-
-		$conf->{$this->config_key} = $extensions;
+		$conf->extensions = $extensions;
 		$conf->save();
 	}
 
@@ -389,7 +377,7 @@ abstract class Minz_Extension {
 
 	public final function saveFile(string $filename, string $content): void {
 		$username = Minz_User::name();
-		$path = USERS_PATH . "/{$username}/{$this->config_key}/{$this->getName()}";
+		$path = USERS_PATH . "/{$username}/extensions/{$this->getName()}";
 
 		if (!file_exists($path)) {
 			mkdir($path, 0777, true);
@@ -400,7 +388,10 @@ abstract class Minz_Extension {
 
 	public final function removeFile(string $filename): void {
 		$username = Minz_User::name();
-		$path = USERS_PATH . "/{$username}/{$this->config_key}/{$this->getName()}/{$filename}";
+		if ($username == null) {
+			return;
+		}
+		$path = USERS_PATH . "/{$username}/extensions/{$this->getName()}/{$filename}";
 
 		if (file_exists($path)) {
 			unlink($path);
